chore(mountpoint): Update fuser to 0.17
- Update include proper types for most elements in `trait FileSystem`
- No longer need to spawn our own thread, `Session::spawn` does that for ourself
- Remove `SessionUnmounter` in favor of `BackgroundSession::umount_and_join`

+++ libparsec/crates/types/src/id.rs
@@ -868,7 +868,7 @@ impl Display for UserProfile {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct FileDescriptor(pub u32);
+pub struct FileDescriptor(pub u64);
 
 #[cfg(test)]
 #[path = "../tests/unit/id.rs"]
